Island logic now has recalculation of connectedness
